feat: pages a propos et mentions legales en markdown
