+++ package.json
@@ -24,6 +24,7 @@
     "babel-loader": "^8.2.2",
     "css-loader": "^5.2.6",
     "express": "^4.17.1",
+    "raw-loader": "^4.0.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
@@ -33,6 +34,11 @@
   },
   "dependencies": {
     "react-data-table-component": "^7.0.0-rc2",
+    "react-markdown": "^8.0.5",
+    "rehype-raw": "^6.1.1",
+    "remark-directive": "^2.0.1",
+    "remark-emoji": "^3.1.0",
+    "remark-gfm": "^3.0.1",
     "styled-components": "^5.3.0",
     "swagger-ui": "^4.1.3"
   }
