+++ src/data/content/about.md
@@ -1,0 +1,29 @@
+[comment]: # (Contenu de la page A propos)
+
+<p class="lead">Cette application permet de tester IGNF/validator-api</p>
+
+## Le principe
+
+Cette application permet de valider et de normaliser les données géographiques et documents présents dans une arborescence de fichiers par rapport à un modèle.
+
+Ce service ne nécessite aucune authentification et ne collecte donc aucune donnée personnelle.
+
+Toutes les validations sont publiques et consultables par quiconque en possède le lien.
+
+Les données ne sont pas conservées au delà d'un délai d'un mois sur le serveur, et ce quel que soit le résultat de la validation qui les concerne.
+
+## Les modèles
+
+### Standards de documents d'urbanisme
+
+Ces standards proviennent du [Géoportail de l'urbanisme](https://www.geoportail-urbanisme.gouv.fr/).
+
+> :warning: **Attention**, les documents validés par ce démonstrateur ne sont pas forcément valides pour le Géoportail de l'urbanisme, car ce dernier réalise des contrôles supplémentaires.
+
+### Standard PCRS
+
+Il s'agit du standard PCRS (**P**lan **C**orps de **R**ue **S**implifié) v2.0 élaboré par le [Groupe de Travail PCRS du CNIG](http://cnig.gouv.fr/gt-pcrs-accompagnement-a1444.html).
+
+### Standard Naviforest
+
+Il s'agit du standard national de données dessertes pour le transport de bois élaboré par le [Groupe de travail du CNIG](http://cnig.gouv.fr/gt-dessertes-pour-les-transports-de-bois-a18535.html).
