+++ src/data/content/legal.md
@@ -1,0 +1,21 @@
+### Editeur
+
+Le site {url} est édité par l’Institut national de l’information géographique et forestière (IGN), 73 avenue de Paris, 94165 SAINT-MANDE CEDEX, France.
+
+Tél : 01 43 98 80 00.
+
+Directeur de la publication : Sébastien Soriano, Directeur général de l'IGN.
+
+### Hébergeur
+
+Le site {url} est hébergé par Cegedim S.A., 137 rue d'Aguesseau, 92641 Boulogne-Billancourt Cedex, France.
+
+### Droit de propriété intellectuelle
+
+Les illustrations, le contenu éditorial et les divers éléments de la charte graphique sont des éléments dont l’IGN détient la propriété ou le droit d’exploitation.
+
+L’utilisation des documents ou éléments du site est soumise à l’accord préalable de notre établissement.
+
+### Liens hypertextes proposés sur ${url}
+
+Les liens insérés dans les pages du site {url} vers des sites tiers sont proposés à titre d’information ; le contenu des sites vers lesquels ces liens conduisent n’engage pas la responsabilité de l’éditeur.
