--- src/data/naviforest-standards.json
@@ -1,6 +1,0 @@
-[
-    {
-        "name": "NAVIFOREST",
-        "url": "https://ignf.github.io/validator-config-naviforest/config/cnig_SNDD_2019/files.json"
-    }
-]
