+++ src/data/standards/naviforest-standards.json
@@ -1,0 +1,6 @@
+[
+    {
+        "name": "NAVIFOREST",
+        "url": "https://ignf.github.io/validator-config-naviforest/config/cnig_SNDD_2019/files.json"
+    }
+]
