+++ src/data/standards/pcrs-standard.json
@@ -1,0 +1,6 @@
+[
+    {
+        "name": "CNIG_PCRS_v2.0",
+        "url": "https://ignf.github.io/validator/validator-core/src/test/resources/config-json/CNIG_PCRS_v2.0/document.json"
+    }
+]
